refactor(tests): add a test utility file for commonly used functions in tests
