+++ hrms/payroll/doctype/employee_benefit_application/test_employee_benefit_application.py
@@ -10,7 +10,6 @@ from frappe.utils import add_days, date_diff, get_year_ending, get_year_start, g
 from erpnext.setup.doctype.employee.test_employee import make_employee
 from erpnext.setup.doctype.holiday_list.test_holiday_list import set_holiday_list
 
-from hrms.hr.doctype.leave_application.test_leave_application import get_first_sunday
 from hrms.hr.utils import get_holiday_dates_for_employee
 from hrms.payroll.doctype.employee_benefit_application.employee_benefit_application import (
 	calculate_lwp,
@@ -24,6 +23,7 @@ from hrms.payroll.doctype.salary_slip.test_salary_slip import (
 )
 from hrms.payroll.doctype.salary_structure.salary_structure import make_salary_slip
 from hrms.payroll.doctype.salary_structure.test_salary_structure import make_salary_structure
+from hrms.tests.test_utils import get_first_sunday
 
 
 class TestEmployeeBenefitApplication(FrappeTestCase):
